# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/users/ban/{user}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Ban a user",
                "description": "Ban a user.",
                "operationId": "0fe02fc3d6a8a69a39bbfe88eaf496a7",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to ban",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User banned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User banned successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}